requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.4
django-filter==22.1
idna==3.4
ipaddress==1.0.23
monero==1.1.1
numpy==1.23.5
Pillow==9.3.0
pycparser==2.21
pycryptodomex==3.16.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
varint==1.0.2
